Rework fuzzer() to use a mirai backend.
This allows to set a timeout so that expensive functions for which the
current argument doesn't cause a crash can be interrupted after 2 seconds.
This means that there is also no more need to skip functions that contain
readline().

This speeds up normal runs (215s to 30s) with 2 daemons, although tests go
from 5s to 41s because there is an overhead during the self-fuzz tests.
Besides, tests don't use many functions in `funs`, which is where execution
can be better parallelised.

Fixes #4.