Run the mirai tasks using a queue to avoid spurious timeouts.
This launches only as many mirais as available daemons, instead of all at
once. This allows timeouts can be applied correctly and avoids later tasks
to be wrongly timed out if some of the earlier ones are slow.

This also allows to parallelize execution over functions, and thus better
exploit the available daemons, as there is no mirai_collect() operation for
each function being fuzzed.

Fuzzing 140 Luminescence functions on 60 inputs takes 7.7s intead of 11.9s
with 2 daemons, and 3.7s instead of 10.8s with 6 daemons.